Update bundle_ro.properties
This is not ready, hence will be a draft. 192f12754967383111c13d3a2ee48fa3af6d27c1 still unfinished. Do not merge as-is. It is unfinished and untested.

This is a rushed submission. I did not want to lose my progress; don't have time to finish it right now.